% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch}
\alias{watch}
\alias{watch_curated}
\alias{watch_buckets}
\title{Watch the new online data library}
\usage{
watch_curated()

watch_buckets()
}
\value{
data frame with summary of curated datasets
}
\description{
Watch the new online data library
}
\examples{
watch_curated()
}
